edit free allow product

+++ database/migrations/2026_07_01_135602_add_taxi_phone_to_orders_table.php
@@ -4,25 +4,25 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-             $table->string('taxi_phone')->nullable()->after('delivery_id');
+            if (!Schema::hasColumn('orders', 'taxi_phone')) {
+                $table->string('taxi_phone')->nullable()->after('delivery_id');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('orders', function (Blueprint $table) {
-              $table->dropColumn('taxi_phone');
+            if (Schema::hasColumn('orders', 'taxi_phone')) {
+                $table->dropColumn('taxi_phone');
+            }
         });
     }
 };
